refactor: move database folder back to src
