+++ src/albums/albums.module.ts
@@ -1,8 +1,7 @@
 import { Module } from '@nestjs/common';
 
-import { AlbumsDatabase } from 'database';
-
 import { ArtistsModule } from '@/artists';
+import { AlbumsDatabase } from '@/database';
 import { EntityName, SharedModule, TOKEN_DATABASE } from '@/shared';
 
 import { AlbumsController } from './albums.controller';
