+++ src/artists/artists.module.ts
@@ -1,7 +1,6 @@
 import { Module } from '@nestjs/common';
 
-import { ArtistsDatabase } from 'database';
-
+import { ArtistsDatabase } from '@/database';
 import { EntityName, SharedModule, TOKEN_DATABASE } from '@/shared';
 
 import { ArtistsController } from './artists.controller';
