+++ src/database/favorites.database.ts
