+++ src/users/users.module.ts
@@ -1,7 +1,6 @@
 import { Module } from '@nestjs/common';
 
-import { UsersDatabase } from 'database';
-
+import { UsersDatabase } from '@/database';
 import { EntityName, SharedModule, TOKEN_DATABASE } from '@/shared';
 
 import { PasswordService } from './password.service';
